% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2plot.R
\name{s2plot}
\alias{s2plot}
\alias{s2plot_par_default}
\title{Plot an object on the sphere}
\usage{
s2plot(
  geog,
  ...,
  projection = s2plot_projection_default(geog, add),
  xlim = NULL,
  ylim = NULL,
  par = s2plot_par_default(),
  add = FALSE
)

s2plot_par_default()
}
\arguments{
\item{geog}{A \code{\link[s2:as_s2_geography]{s2::as_s2_geography()}}}

\item{...}{Passed to graphics functions.}

\item{projection}{Right now \code{\link[=s2plot_projection_orthographic]{s2plot_projection_orthographic()}}
is the only projection. The default is either calculated
based on the \code{\link[s2:s2_boundary]{s2::s2_centroid_agg()}} of \code{geog} or the
last used projection if \code{add = TRUE}.}

\item{xlim, ylim}{Limits in projected space.}

\item{par}{Graphical \code{\link[graphics:par]{graphics::par()}} to set prior to plotting}

\item{add}{Add to the current plot? Use \code{FALSE} to create a new plot.}
}
\value{
\code{x}, invisibly
}
\description{
Plot an object on the sphere
}
\examples{
s2plot(s2::s2_data_countries())
s2plot(s2::s2_data_cities("London"), pch = 16, add = TRUE)

}
