% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{s2plot_projection_orthographic}
\alias{s2plot_projection_orthographic}
\alias{s2plot_projection_default}
\alias{s2plot_prepare}
\alias{s2plot_project}
\alias{s2plot_prepare.s2plot_projection_orthographic}
\alias{s2plot_project.s2plot_projection_orthographic}
\title{Define a projection to use with s2plot}
\usage{
s2plot_projection_orthographic(point, rotation = 0)

s2plot_projection_default(geog, add = FALSE)

s2plot_prepare(projection, geog)

s2plot_project(projection, geog)

\method{s2plot_prepare}{s2plot_projection_orthographic}(projection, geog)

\method{s2plot_project}{s2plot_projection_orthographic}(projection, geog)
}
\arguments{
\item{point}{The centre of the projection}

\item{rotation}{Rotation}

\item{geog}{A \code{\link[s2:as_s2_geography]{s2::as_s2_geography()}}}

\item{add}{Add to the current plot? Use \code{FALSE} to create a new plot.}

\item{projection}{Right now \code{\link[=s2plot_projection_orthographic]{s2plot_projection_orthographic()}}
is the only projection. The default is either calculated
based on the \code{\link[s2:s2_boundary]{s2::s2_centroid_agg()}} of \code{geog} or the
last used projection if \code{add = TRUE}.}
}
\description{
Define a projection to use with s2plot
}
