% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2plot-package.R
\docType{package}
\name{s2plot-package}
\title{s2plot: Plot spatial objects on a sphere}
\description{
Experimental spherical plotting of spatial objects.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
