% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2plot.R
\name{s2plot}
\alias{s2plot}
\alias{s2plot_new}
\alias{s2plot_draw}
\title{Plot an object on the sphere}
\usage{
s2plot(geog, ..., xlim = NULL, ylim = NULL)

s2plot_new(geog, ..., xlim = NULL, ylim = NULL)

s2plot_draw(geog, ...)
}
\arguments{
\item{geog}{A \code{\link[libs2:s2geography]{libs2::s2geography()}}}

\item{...}{Passed to graphics functions.}

\item{xlim, ylim}{Limits in projected space (usually in the range -1 to 1).}
}
\value{
\code{x}, invisibly
}
\description{
Plot an object on the sphere
}
