% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2plot.R
\name{s2plot}
\alias{s2plot}
\alias{s2plot_par_default}
\title{Plot an object on the sphere}
\usage{
s2plot(
  geog,
  ...,
  projection = s2plot_projection_default(geog),
  xlim = NULL,
  ylim = NULL,
  par = s2plot_par_default(),
  add = FALSE
)

s2plot_par_default()
}
\arguments{
\item{geog}{A \code{\link[libs2:s2geography]{libs2::s2geography()}}}

\item{...}{Passed to graphics functions.}

\item{xlim, ylim}{Limits in projected space.}
}
\value{
\code{x}, invisibly
}
\description{
Plot an object on the sphere
}
