% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{s2plot_projection_orthographic}
\alias{s2plot_projection_orthographic}
\alias{s2plot_projection_default}
\alias{s2plot_prepare}
\alias{s2plot_project}
\alias{s2plot_prepare.s2plot_projection_orthographic}
\alias{s2plot_project.s2plot_projection_orthographic}
\title{Define a projection to use with s2plot}
\usage{
s2plot_projection_orthographic(point, rotation = 0)

s2plot_projection_default(geog)

s2plot_prepare(projection, geog)

s2plot_project(projection, geog)

\method{s2plot_prepare}{s2plot_projection_orthographic}(projection, geog)

\method{s2plot_project}{s2plot_projection_orthographic}(projection, geog)
}
\arguments{
\item{point}{The centre of the projection}

\item{rotation}{Rotation}

\item{geog}{A \code{\link[libs2:s2geography]{libs2::s2geography()}}}

\item{projection}{Right now \code{\link[=s2plot_projection_orthographic]{s2plot_projection_orthographic()}}
is the only projection.}
}
\description{
Define a projection to use with s2plot
}
